+++ src/MultipleInput.php
@@ -74,7 +74,7 @@ class MultipleInput extends InputWidget
     public $min;
 
     /**
-     * @var string position of add button. By default button is rendered in the row.
+     * @var string|array position of add button. By default button is rendered in the row.
      */
     public $addButtonPosition = self::POS_ROW;