+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Yii2 multiple input change log
 - #79 Added support for embedded MultipleInput widget (unclead, execut)
 - Enh: Added ability to render `add` button in the footer (unclead)
 - Enh: Improving for better work without ActiveForm (unclead)
+- Enh: Added ability to render `add` button at several positions (unclead)
 
 1.2.19
 ------